requirements.txt
asgiref==3.7.2
Django==4.2.11
django-filter==24.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
install==1.3.5
Markdown==3.6
mongoengine==0.28.2
mysqlclient==2.2.4
psycopg2==2.9.9
pymongo==3.12.3
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
